// end-to-end-solutions/Luna/src/Luna.UI/isv_client/src/models/IOfferParameterModel.ts
import {IBaseModel} from "./IBaseModel";

export interface IOfferParameterModel extends IBaseModel {
  parameterName: string;
  displayName: string;
  description: string;
  valueType: string;
  fromList: boolean;
  valueList: string;
  maximum: number | null;
  minimum: number | null;
  isdisablemaximum:boolean;
  isdisableminimum:boolean;
  isdisablevalueList:boolean;
  isdisablefromList:boolean;  
}